fix tests, add more tests for counter
