+++ codejam/client/widgets/draw_canvas.py
